requirements.txt
fastapi
uvicorn
torch
torchvision
matplotlib
numpy
Pillow
